+++ src/main/java/io/electrum/moneytransfer/api/IAdminResource.java
@@ -65,6 +65,19 @@ public interface IAdminResource {
          HttpServletRequest httpServletRequest)
          throws NotFoundException;
 
+   @Deprecated
+   void getExchangeRate(
+         @NotNull String fromCurrency,
+         String toCurrency,
+         String receiverId,
+         SecurityContext securityContext,
+         Request request,
+         HttpHeaders httpHeaders,
+         AsyncResponse asyncResponse,
+         UriInfo uriInfo,
+         HttpServletRequest httpServletRequest)
+         throws NotFoundException;
+
    void getExchangeRate(
          @NotNull String fromCurrency,
          String toCurrency,