+++ src/main/java/io/electrum/moneytransfer/api/IOrdersResource.java
@@ -57,6 +57,20 @@ public interface IOrdersResource {
          UriInfo uriInfo,
          HttpServletRequest httpServletRequest);
 
+   @Deprecated
+   void lookupOrder(
+         String orderRedeemRef,
+         String merchantId,
+         @NotNull String originatorInstId,
+         @NotNull String receiverId,
+         String remittanceRef,
+         SecurityContext securityContext,
+         Request request,
+         HttpHeaders httpHeaders,
+         AsyncResponse asyncResponse,
+         UriInfo uriInfo,
+         HttpServletRequest httpServletRequest);
+
    void lookupOrder(
          String orderRedeemRef,
          String merchantId,